feat(app): update routes to new format
BREAKING CHANGE: This change introduces a new way of scoping routes.
